Large commit, new circlebuffer now correctly works
Many fixes in mutiple places

+++ Utility/Utilities/Include/PointerArithmetic/PointerArithmetic.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 namespace Doremi
 {
     namespace Utilities
@@ -16,9 +18,9 @@ namespace Doremi
 
             // TODORT
             // Not sure how this works if one attempts to use "real types" instead of pointers.
-            template <typename T> size_t Difference(T const p_low, T const p_high)
+            template <typename T> intmax_t Difference(T const p_low, T const p_high)
             {
-                return (reinterpret_cast<size_t>(p_high) - reinterpret_cast<size_t>(p_low));
+                return (reinterpret_cast<intmax_t >(p_high) - reinterpret_cast<intmax_t >(p_low));
             }
 
             static bool AssertAdresstInside(void* const p_adressToAssert, void* const p_low, void* const p_high)
